+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - Record Deletions - Handle the record deletion on models having a composite primary key.
 - Liana Version & Orm Version - Prevent server crash on bad format version.
+- Apimaps - Prevent foreign key field creation for belongsTo associations having a constaints set to false.
 
 ## RELEASE 1.1.9 - 2017-06-28
 ### Fixed